configs: Make SLMK less aggresive
Set SLMK min free to 125MB that background apps and tasks are less likely to be killed to free up memory.
This can allow more apps to stay active in memory, making it easier to switch between them without needing to reload or restart them.

As a result, multitasking can feel smoother and more efficient

Signed-off-by: Santhosh <santhosh.user.why.red@gmail.com>